fix: notification mark as read error for broadcast notifications
- Fix route model binding issue by using plain ID parameter
- Allow marking broadcast notifications (notifiable_type='all') as read
- Update all notification controllers (User, Dosen, Admin)
- Update routes to use {id} instead of {notification}

+++ app/Http/Controllers/Admin/AdminNotificationController.php
@@ -10,8 +10,11 @@ class AdminNotificationController extends Controller
 {
     public function markAsRead($id)
     {
-        $notification = AppNotification::findOrFail($id);
-        $notification->markAsRead();
+        $notification = AppNotification::find($id);
+        
+        if ($notification) {
+            $notification->markAsRead();
+        }
         
         return back();
     }
@@ -29,8 +32,11 @@ class AdminNotificationController extends Controller
 
     public function destroy($id)
     {
-        $notification = AppNotification::findOrFail($id);
-        $notification->delete();
+        $notification = AppNotification::find($id);
+        
+        if ($notification) {
+            $notification->delete();
+        }
         
         return back();
     }
